requirements.txt
Flask==1.1.1
python-telegram-bot==12.4.2
gunicorn==20.0.4
requests==2.23.0
